moved the encoding and the interpreter classes into the encoding project
