+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/constraints/ElementList.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *******************************************************************************/
-package net.sf.opendse.optimization.constraints;
+package net.sf.opendse.encoding.constraints;
 
 import java.util.ArrayList;
 import java.util.Collection;
