+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/constraints/SpecificationCapacityConstraints.java
@@ -19,12 +19,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *******************************************************************************/
-package net.sf.opendse.optimization.constraints;
+package net.sf.opendse.encoding.constraints;
 
+import static net.sf.opendse.encoding.old.variables.Variables.p;
+import static net.sf.opendse.encoding.old.variables.Variables.var;
 import static net.sf.opendse.model.Models.filterCommunications;
 import static net.sf.opendse.model.Models.getLinks;
-import static net.sf.opendse.optimization.encoding.variables.Variables.p;
-import static net.sf.opendse.optimization.encoding.variables.Variables.var;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -44,9 +44,10 @@ import net.sf.opendse.model.Mappings;
 import net.sf.opendse.model.Resource;
 import net.sf.opendse.model.Routings;
 import net.sf.opendse.model.Specification;
+import net.sf.opendse.model.SpecificationWrapper;
 import net.sf.opendse.model.Task;
 import net.sf.opendse.model.Models.DirectedLink;
-import net.sf.opendse.optimization.SpecificationWrapper;
+
 
 import org.opt4j.satdecoding.Constraint;
 import org.opt4j.satdecoding.Constraint.Operator;
