+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/constraints/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package for the classes related to the
+ * {@link org.opt4j.satdecoding.Constraint}s used for the encoding of valid
+ * implementations.
+ */
+package net.sf.opendse.encoding.constraints;
