+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/old/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package containing the classes used for the implementation encoding not
+ * following the modular approach.
+ * 
+ */
+package net.sf.opendse.encoding.old;
