+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/old/variables/CLRRP.java
@@ -19,40 +19,41 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *******************************************************************************/
-package net.sf.opendse.optimization.constraints;
+package net.sf.opendse.encoding.old.variables;
 
-import java.util.Collection;
-import java.util.Set;
+import net.sf.opendse.model.Edge;
+import net.sf.opendse.model.ICommunication;
+import net.sf.opendse.model.Resource;
+import net.sf.opendse.model.Task;
 
-import net.sf.opendse.model.Specification;
-import net.sf.opendse.model.parameter.ParameterReference;
+public class CLRRP extends Variable implements CommunicationVariable {
 
-import org.opt4j.satdecoding.Constraint;
-import org.opt4j.satdecoding.Model;
+	public CLRRP(Task t, Edge l, Resource r0, Resource r1, Task p) {
+		super(t,l,r0,r1,p);
+	}
 
-public interface SpecificationConstraints {
-	
-	public void doEncoding(Collection<Constraint> constraints);
-	
-	public void doInterpreting(Specification implementation, Model model);
-	
-	public Set<ParameterReference> getActiveParameters();
+	public Task getTask() {
+		return get(0);
+	}
 
-	public static String CONNECT = ":CONNECT";
-	public static String CONNECT_MAX = ":CONNECT-MAX";
-	public static String CONNECT_MIN = ":CONNECT-MIN";
-	
-	public static String CAPACITY = ":CAPACITY";
-	public static String CAPACITY_MAX = ":CAPACITY-MAX";
-	public static String CAPACITY_MIN = ":CAPACITY-MIN";
-	public static String CAPACITY_SCALE = ":CAPACITY-SCALE";
-	public static String CAPACITY_ACTION = ":CAPACITY-ACTION";
-	public static String CAPACITY_VALUE = ":CAPACITY-VALUE";
-	public static String CAPACITY_RATIO = ":CAPACITY-RATIO";
-	
-	public static String ELEMENTS_EXCLUDE = "ELEMENTS_EXCLUDE";
-	public static String ELEMENTS_REQUIRE = "ELEMENTS_REQUIRE";
-	
-	public static String ROUTER = "ROUTER";
+	public Edge getLink() {
+		return get(1);
+	}
 
+	public Resource getSource() {
+		return get(2);
+	}
+
+	public Resource getDest() {
+		return get(3);
+	}
+	
+	public Task getP(){
+		return get(4);
+	}
+	
+	@Override
+	public ICommunication getCommunication() {
+		return (ICommunication)getTask();
+	}
 }
