+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/old/variables/CR.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *******************************************************************************/
-package net.sf.opendse.optimization.encoding.variables;
+package net.sf.opendse.encoding.old.variables;
 
 import net.sf.opendse.model.ICommunication;
 import net.sf.opendse.model.Resource;
