--- opendse-encoding/src/test/java/net/sf/opendse/encoding/DummyTestFile.java
@@ -1,20 +1,0 @@
-package net.sf.opendse.encoding;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-/**
- * To be removed; Just there to commit the folder structure.
- * 
- * @author Fedor Smirnov
- *
- */
-public class DummyTestFile {
-
-	@Test
-	public void test() {
-		assertTrue(true);
-	}
-
-}
