+++ opendse-optimization/src/main/java/net/sf/opendse/optimization/DesignSpaceExplorationDecoder.java
@@ -31,6 +31,7 @@ import net.sf.opendse.model.Attributes;
 import net.sf.opendse.model.Element;
 import net.sf.opendse.model.Models;
 import net.sf.opendse.model.Specification;
+import net.sf.opendse.model.SpecificationWrapper;
 import net.sf.opendse.model.parameter.Parameter;
 import net.sf.opendse.model.parameter.ParameterReference;
 import net.sf.opendse.model.parameter.ParameterSelect;
