+++ opendse-optimization/src/main/java/net/sf/opendse/optimization/RoutingVariableClassOrder.java
@@ -2,12 +2,12 @@ package net.sf.opendse.optimization;
 
 import com.google.inject.Inject;
 
+import net.sf.opendse.encoding.old.Encoding;
+import net.sf.opendse.encoding.old.variables.CR;
+import net.sf.opendse.encoding.old.variables.EAVI;
 import net.sf.opendse.model.Link;
 import net.sf.opendse.model.Mapping;
 import net.sf.opendse.model.Resource;
-import net.sf.opendse.optimization.encoding.Encoding;
-import net.sf.opendse.optimization.encoding.variables.CR;
-import net.sf.opendse.optimization.encoding.variables.EAVI;
 
 /**
  * When bound during an exploration, the {@link RoutingVariableClassOrder}
