+++ opendse-optimization/src/main/java/net/sf/opendse/optimization/SATCreatorDecoder.java
@@ -27,9 +27,11 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
+import net.sf.opendse.encoding.ImplementationInterpreter;
 import net.sf.opendse.model.Resource;
 import net.sf.opendse.model.Specification;
-import net.sf.opendse.optimization.encoding.Interpreter;
+import net.sf.opendse.model.SpecificationWrapper;
+
 import org.opt4j.core.Genotype;
 import org.opt4j.core.common.random.Rand;
 import org.opt4j.core.optimizer.Control;
@@ -51,12 +53,12 @@ public class SATCreatorDecoder extends AbstractSATDecoder<Genotype, Implementati
 
 	protected final SATConstraints constraints;
 	protected final SpecificationWrapper specificationWrapper;
-	protected final Interpreter interpreter;
+	protected final ImplementationInterpreter interpreter;
 	protected final Control control;
 
 	@Inject
 	public SATCreatorDecoder(VariableClassOrder order, SATManager manager, Rand random, SATConstraints constraints, SpecificationWrapper specificationWrapper,
-			Interpreter interpreter, Control control,
+			ImplementationInterpreter interpreter, Control control,
 			@Constant(value = "variableorder", namespace = SATCreatorDecoder.class) boolean useVariableOrder) {
 		super(manager, random);
 		this.order = order;
