+++ opendse-optimization/src/test/java/net/sf/opendse/optimization/test/FirstSpecificationTest.java
@@ -21,6 +21,7 @@
  *******************************************************************************/
 package net.sf.opendse.optimization.test;
 
+import net.sf.opendse.encoding.constraints.SpecificationConstraints;
 import net.sf.opendse.io.SpecificationWriter;
 import net.sf.opendse.model.Application;
 import net.sf.opendse.model.Architecture;
@@ -34,7 +35,6 @@ import net.sf.opendse.model.Routings;
 import net.sf.opendse.model.Specification;
 import net.sf.opendse.model.Task;
 import net.sf.opendse.model.parameter.Parameters;
-import net.sf.opendse.optimization.constraints.SpecificationConstraints;
 import net.sf.opendse.visualization.SpecificationViewer;
 
 public class FirstSpecificationTest {
