+++ opendse-tutorial/src/main/java/net/sf/opendse/tutorial/Part4.java
@@ -1,5 +1,6 @@
 package net.sf.opendse.tutorial;
 
+import net.sf.opendse.encoding.constraints.SpecificationConstraints;
 import net.sf.opendse.io.SpecificationWriter;
 import net.sf.opendse.model.Application;
 import net.sf.opendse.model.Architecture;
@@ -11,7 +12,6 @@ import net.sf.opendse.model.Mappings;
 import net.sf.opendse.model.Resource;
 import net.sf.opendse.model.Specification;
 import net.sf.opendse.model.Task;
-import net.sf.opendse.optimization.constraints.SpecificationConstraints;
 import net.sf.opendse.optimization.encoding.SingleImplementation;
 import net.sf.opendse.visualization.SpecificationViewer;
